requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.2.10
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
blinker==1.4
Brlapi==0.7.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==3.0
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.1.15
django-crispy-forms==1.9.2
django-pyodbc-azure==2.1.0.0
django-storages==1.9.1
djangorestframework==3.11.1
entrypoints==0.3
filelock==3.0.12
httplib2==0.14.0
idna==2.10
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
mysqlclient==1.4.4
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyodbc==4.0.30
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-decouple==3.3
python-dotenv==0.14.0
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.24.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
sqlparse==0.3.1
system-service==0.3
systemd-python==234
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.10
virtualenv==20.0.28
wadllib==1.3.3
xkit==0.0.0
